requirements.txt
loguru==0.5.3
lxml==4.6.3
numpy==1.20.2
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
xlrd==2.0.1
